Refactor serializers, replace bibdata w/discover
This is the big refactor that implements `discover` cores in place of
`bibdata` for API resources, mainly `bib`. In the process of working
through this I also decided to tackle refactoring our simpleserializers
and related classes to make them a little more flexible. The goal was to
let us more easily implement API fields that are a combination of Solr
fields (sorting, searching, and filtering are handled by separate Solr
fields).

Notes about a few other changes in this commit:

- Item, EResource, and Bib record IDs are the III record number instead
of the int ID.

- Record numbers now do NOT include the III check digit, to match
Discover record numbers/IDs.

- (Record number is still a separate field on the resource, which
duplicates the ID, for compatibility.)

- Changed author_sort field from norm_string to str. In the discover
schemas, there is no need for the author_sort field to be `norm_string`
type -- it is normalized during the indexing pipeline and can be stored
as a Solr string type, like the other sort fields. (This way we can use
docValues with it.)