Improve performance of new serializer code
I noticed that tests involving the refactored serializer code were much
slower following the refactoring. So, I did some timing of the API tests,
comparing multiple versions, to try to determine whether the poorer
performance is a result of the new Solr version, using the `discover`
indexes in place of `bibdata` for the `bib` resource, or the new
serializer code. The answer is: yes.

This commit should at least fix the new serializer code; it now performs
at least as well (maybe slightly better) compared to the old code. Before
this change, it was forcing every single field to go through the
`cast_to_python` process even though 90% of them didn't need to because
they're just string fields. I added a flag to the `SimpleField` classes
that lets you bypass `cast_to_python` for field types (like numbers and
strings) that work fine without it. The amount of time to serialize a
list of objects is ~1/4 what it was before this change.

We can work on benchmarking Solr separately and hopefully address that
as well, later.