Fix (longstanding?) issue with serializer caching
While testing after routine server OS updates I noticed that, when
trying to use the inventory app, sometimes saving changes results in an
error where the server seems to believe that I'm trying to change a
non-writeable field (specifically `status`). I traced the problem back
to the `sierra.api.simpleserializer.SimpleSerializerWithLookups` class.
For some reason, when I get the error, it believes that the `status`
field coming from Solr is blank and I'm trying to overwrite it. (All
other fields are populated.) The `status` field comes from a lookup, and
I believe the problem was that whatever HTTP process was handling the
bad requests had NOT cached this lookup at the time of the requests, so
it was (erroneously) returning an empty `status` value. I seem to have a
vague memory of this happening before -- this happens if/when the
inventory app doesn't get enough use to keep that cache populated for
all the active HTTP handlers. Refreshing the page several times is
usually enough to fix it. However, I took another look at the code and
implemented a simple bandaid fix for it: now, if a lookup fails once,
before returning a blank value it will try to refresh the lookup cache
from Solr. This should prevent this problem from recurring.