Add auto-notes feature to inventory util
Librarians at Sycamore want to use the inventory app to do another
inventory of their materials, and they want to reset their items. This
is what the inventory util functions are for. However, they don't want
to clear out everything, they want to keep existing inventory notes. To
do this, we have to be able to auto-generate new system notes that
reflect the actions that are taken. If an item is set "on shelf" and we
clear that status, we have to generate a system note saying that we
cleared it. Otherwise, the inventory app will consider the system notes
to be invalid when you try to change the shelf status via the app,
later.

This change adds that capability. Now, when using `set_item_fields`, you
can pass in the `auto_notes=True` argument. If the values you want to
set don't include inventory_notes and they do include shelf_status or
flags, then it will automatically add new system-generated notes to any
existing inventory notes.