Add default ALLOWED_HOSTS to dev settings
When I attempted spinning up the app in a dev environment and
connecting to it, it complained that 'localhost' needed to be added to
ALLOWED_HOSTS before I could access it. This must have been a default
that was removed at some point by Django. For dev settings, 'localhost'
and '127.0.0.1' both are now added explicitly.