Fix ResourceTypeDeterminer.get_callnums_from_obj
In an active dev environment, I tried loading a batch of ~60,000 bibs +
attached items. The only issue I saw was ~50 or so records that couldn't
be loaded because the ResourceTypeDeterminer choked with an unspecified
TypeError. Upon investigating, I traced the problem to the
'get_callnums_fom_obj' method -- specifically, when it tried to order
item call numbers based on the 'items_display_order' property from
bibrecorditemrecordlinks, if that property is None, then it raised a
"'<' not supported between instances of 'NoneType' and 'NoneType'"
error. (This is something that Python 3 complains about that Python 2
allowed.) There is a weird elusive Sierra issue that crops up in a small
number of records where 'items_display_order' is None, so I couldn't
leave this unaddressed.

To fix it, I modified the sort key so that it sorts these here the way
Sierra actually sorts them -- first by 'items_display_order' and then
(if that's None) by item record number.

I also added a test for this.