Fix SimpleSerializerWithLookups `None` lookup
Tiny change, but one that has a decent impact in testing. When using a
REST API serializer that inherits from the SimpleSerializerWithLookups
class, and you use the 'get_lookup_value' to look up a code in the
cache -- if your lookup code is None, now the lookup will simply return
None instead of treating it as a cache miss and refreshing the cache.
(Refreshing the cache of course fires another lookup query, hitting your
data store in a case where you wouldn't expect any return value.) So,
this change helps prevent some unnecessary queries in these situations.