Add replication configuration to solrconfig.xml
In getting this ready to be deployed via Ansible for the libui03 stack,
I've added ReplicationHandler configuration to solrconfig.xml for all
three Solr cores. In production we're moving to a multi-server setup
that uses replication to distribute copies of the index between nodes.
Since this is only for production, the new ReplicationHandler
configuration is commented out. With Ansible, we'll use a series of
regex replace tasks to remove or place comment tags immediately
surrounding each requestHandler block, depending on whether a specific
host is a leader, follower, or neither.

I'm making the same update to the solrconfig for the 'libguides' core in
the 'libguide-extract' project.