Log to the console when using 'dev' settings
Deploying this in a more production-like environment, I'm having to do a
bit of debugging using the app. I noticed it no longer logs INFO-level
messages to the console by default, which is indeed how the loggers are
set up, but I'm not sure when/why we made that change. In the 'dev'
settings, I've added an override for the base 'django' logger that
changes this back so it works as expected in dev.