Fix SimpleSerializerWithLookups 1-resource lookup
Ran into an obscure error when we deployed this out to an
Apache/mod_wsgi site, where sometimes hitting a resource the uses the
SimpleSerializerWithLookups serializer without first hitting the list
view it appears in raises an error. Found that I'd made a simple mistake
with a 'getattr' lookup that caused it to error when the
'get_lookup_value' method triggers a 'cache_all_lookups' refresh.

I didn't have any tests that tested this situation, so I added one. I
confirmed that it raised the error in question and that fixing the error
causes the test to pass.