Fix haystack's use of id, django_id, haystack_id
One of the changes I've made in refactoring this past ~year has been to
change the Solr/API "id" fields for items, bibs, and eresources to use
the III record number instead of the internal Sierra PK. But apparently
I didn't fully update things -- particularly the shelflist app code and
test data factories/profiles for Solr -- to account for how this would
play in the haystack index. The haystack index contains records of
various types, not just items, and so it uses various ID fields for
varying degrees of uniqueness; the 'haystack_id' field is the fully
qualified ID (used as the "unique" field in the Solr schema), while the
'django_id' (internal), 'id' (public ID for items/eresource), and 'code'
(public ID for locations/etc.) fields may not be unique.

When I switched to using the III record number in the 'id' field, I
failed to update the various custom lookup methods on the
ShelflistItemIndex class. Previously I had hardcoded these to do lookups
against the Solr 'id' field using the Sierra record PK. This introduced
a couple of bugs.

1. The lookup that happens when you run the shelflistitems exporter and
   reindex item records from Sierra into Solr failed to match anything
   because it was looking up the wrong IDs. So it was unable to
   find/merge user-created fields with fields from the Sierra record.
2. Code related to generating the shelflist-item manifests did not
   always work correctly because it sometimes created lookups based on
   the wrong ID values.

To fix these, I changed the lookups to use the fully qualified
'haystack_id' field in conjunction with the 'get_qualified_id' method on
the indexer. This is the more standard way to get these ID values and
should help future-proof this code.

Our tests *should* have caught these problems, but apparently I did not
update them completely or correctly when I originally tried to switch
from the old-style to the new-style 'id' values. The test data that was
loaded into Solr still used the old-style 'id' values, and for some
reason I had modified the tests so that they expected newly loaded data
to match on the old 'id' and then update the record to use the new 'id'.
But this is not how it works. I've updated the tests so that they now
load test data using the 'haystack_id' field and the 'get_qualified_id'
method. Since the test data now matches what the code under test
expects, the tests are now testing the correct thing.

While I was at it, I updated two additional things in tests:

1. I edited the 'itemstosolr_shelflist_manifests' test to make a clear
   distinction between the 'django_id' and 'id' fields, and to pattern
   each appropriately. Record 'id's are loaded into Redis in the
   predefined shelflist order. No changes to anything but the test were
   required, it just helped clarify what's happening there now.
2. I updated our solr test profiles to ensure they match exactly how the
   'id', 'django_id', and 'haystack_id' fields actually interact.
   Previously I did have it set to mock a III record number in the 'id'
   field, but (weirdly) I had 'django_id' set to copy the value from
   'id'. Now I've simply updated the 'django_id' generators to create an
   auto-increment number starting at (e.g.) 1. All tests still pass, so
   it didn't really seem to *matter* -- but at least the test data we
   generate now better matches the data we derive from Sierra records.