Fix bib->item link URLs in api
While looking at this on one of the new libui03 servers I noticed the
links to item records displayed via the 'bibs' resource were missing the
leading 'i' in the item IDs. (E.g., instead of i000001 they just
contained 0000001.) I traced this back to the fact that we're now
pulling this data from the 'items_json' Discover field, and the item IDs
in that data structure (the 'i' key) do not have the leading 'i'.

We weren't catching this in tests because the test Solr profile which
generated the data used in tests DID include the leading 'i'.

I've fixed this both in the test Solr profile and in the BibSerializer.