Ensure SimpleDateTimeField handles null values
I'm FINALLY deploying the code that was refactored over the last couple
of years. In changes around api.simpleserializers -- I made them a
little more complex by implementing field classes, to help improve how
specific types like datetimes are handled. Apparently in the
SimpleDateTimeField conversion from a Python datetime object to a value
to send to Solr, I didn't account for the possibility of a null value
(None) from the client -- such as, the client wishes to delete an
existing datetime and passes a null value. It was causing an error when
trying to convert None into a datetime string.

The fix was trivial. I also added a test that tests this edge case.