Strip URI from author names (ItemIndex)
A few years ago we began including $0 with URIs in our bib records,
usually for authors and other name fields. I hadn't realized until now
that the "best_author" field we pull from Sierra, which we use in
base.search_indexes.ItemIndex for the 'parent_bib_main_author' field,
includes the $0. (It displays in the inventory app.) I've added a line
to strip out any URIs from this field.

Future TODO: utilize the export.marcparse code to generate values for
the 'parent_bib_*' fields that match what appear in the 'bib' resource.