Add new check for languages from title (for facet)
This addresses issue 259 in the Discover repository. We have noticed
some garbage that appears in the language facet, which appears to come
from the $l in uniform title fields. I've added an extra check to the
pipeline that happens when dumping language data from the uniform titles
into the language facet: if the language value is not one that appears
in the list of MARC languages, then ignore it.