Fix issue w/WordDelimiter filter in Solr schemas
After upgrading to Solr 9, we ran across an issue in production where
long (and even shorter) citation-style queries were triggering errors
saying the maxBooleanClauses limit had been exceeded, which is 1024 by
default. I traced the problem back to the query-expansion that happens
with punctuation, specifically the WordDelimiterGraphFilterFactory
class. I tried raising the maxBooleanClauses limit, but I found I had to
raise it by a factor of 10 to avoid the error even for shorter
citations. And, even when queries completed, they took 7-8 seconds to
run.

After some experimentation, I found that setting `preserveOriginal` to
0 fixed the problem: the same query that took 7-8 seconds and required
raising the maxBooleanClauses limit to 10240 ran in ~315ms, under the
original 1024 limit.

This commit changes that setting for all "text" fields where that
applies. The one field type I've left as-is is the "heading_terms_text"
type, because it deals with this differently and requires commas to be
kept as individual terms.