Add utility for generating shelflistitem manifests
Sometimes, rarely, you may find a need to regenerate shelflistitem
manifests (in Redis) without reindexing anything in Solr. This adds a
utility to `utils.inventory` that does this for each location in a list
of locations.

This also adds a simple `get_locations` function that gets the full list
of locations for you.