Accommodate setting / getting empty vals in Redis
When I deployed the last batch of changes to staging and tested, the
EResourcesToSolr process triggered a problem with the refactored
utils.redisobjs code that I had not anticipated. When it generates
holdings for each eresource and attempts to save the applicable
lists/mappings to Redis, a resource with no holdings tries to save an
empty holdings list. It ends up calling ZADD without arguments, which
causes a Redis error. (Before the refactor, it would loop through a
provided list and set values element by element. For an empty list, no
values would be set.)

When I went back and looked at the redisobjs tests, I realized I had not
at all tested for setting empty and nonexistent values. This adds the
appropriate tests along with some minor changes to make everything work.

Now when you set a Redis key to None or an empty value ('', [], {},
tuple(), set()) then it deletes that key. Getting an unset key returns
None, so the end result is no different than before. Using 'set_field'
with an empty mapping or 'set_value' with an empty values list makes no
changes to the Redis data. When used with a pipeline, the return value
for those operations is None. I also added accommodations for what
should be edge cases where you might call obj.get_field() or
obj.get_index() (i.e., with no arguments). In these cases it just
returns None.