Accommodate multiple dependent deferred updates
While working on adding capabilities to stream batch updates to Redis, I
realized that multiple update operations like appending items to the end
of a list *fail* when RedisObject.defer is True. This is because each
operation extends the length of the list without committing the changes
to Redis. Each update operation therefore tries to set the data at the
exact same index position.

This fixes that by adding the ability to store information about pending
changes in the pipeline object. When updating a list, zset, or string,
it gets the new length for that object there before it tries to get it
from the data as stored in Redis. After the update, it stores the new
length there. And, the pending information is cleared when
`obj.pipe.execute` is called.

I've also changed/clarified how zsets work when setting values at
specific index positions. Index positions are stored directly as
'scores' even when they are out of range -- they are no longer converted
so they append to the very end of the list. This is the most consistent
and straightforward way to do it, instead of trying to manage scores
manually to keep them sequential. Let the client do that if that's what
they need.