Improve Redis encoder/decoder classes
This just does a little minor refactoring of the _RedisJsonEncoder and
_RedisJsonDecoder classes (mainly encoder) to simplify and add some new
functionality.

Encoder methods now return generators so they're easier to use if you
need to encode data separately, apart from the RedisObject.set method,
without making you loop through an entire data set multiple times.

Converts both classes to ones that should be instantiated, where a
'member' property or method is used to encode/decode data in lists,
hashes, zsets, etc. This cuts down on some boilerplate code. It also
lets us use 'ujson.dumps' as the 'member' encoding property directly,
instead of wrapping it in a method.

Removes the 'int' type from _RedisJsonDecoder, since this is no longer
used. (The 'encoded_obj' method takes care of that.)

The _RedisSetter class now allows setting a 'bypass_encoding' property.
Setting this to True indicates that you've already encoded the data from
the user and it can bypass that step when it sends it to Redis. (To
implement this, I also created an 'encode' method to wrap the
'encoder.*' methods. When 'bypass_encoding' is True, it returns the data
as-is without encoding it.)