Refactor _RedisSetter to align with _RedisGetter
This is a small refactor commit that makes changes to _RedisSetter like
the ones the last commit made to _RedisGetter. Most of the logic that
processes arguments and determines what to update and how has been moved
from RedisObject.set into _RedisSetter. Its main method is now
'add_to_pipe' (to add the needed commands to the object's pipeline),
mirroring _RedisGetter.add_to_pipe.

The other change to RedisObject is to add a property setter method for
RedisObject.rtype, which provides a public interface for setting that
property when it's not being pulled from Redis.

Otherwise, most of the changes here are just to help make some sense out
of the previously incomprehensible code that calculates offsets and such
for updates to lists, zsets, and strings based on the 'index' value that
the user provides.