Add RedisObjectStream class for batch operations
When you have a very large amount of data stored in a single Redis key,
it occurred to me that you may want an easy way to break the data into
batches so that you don't saturate the network interface or tie up Redis
for too long at any given time.

This is an initial stab at adding that, but there's still a lot of work
to do on it.