Add __len__ and __bool__ methods
It's useful to know how many pending commands are on a Pipeline object;
this adds a __len__ method for that. A __bool__ method is added so
Pipeline instances will evaluate to True even if they have no pending
commands. (I added a __bool__ method to RedisObject also, since it has a
__len__ method.)