Make shelflist exporters use batch mode w/Redis
When running e.g. ItemsToSolr exports that generate shelflists and store
them in Redis, some of those shelflists can get quite large. This change
simply adds batch options to the code that saves the shelflists in
Redis, taking advantage of the changes to 'utils.redisobjs' -- so we get
the performance improvements from the optimizations without potentially
committing GIANT shelflists to Redis all at once and blocking for
seconds at a time while a transaction runs.