Fix timestamp format in simpleauth
Testing the recent changes to utils.redisobjs on staging, attempting to
log into the inventory app resulted in a TypeError when comparing the
stored (Redis) last login timestamp with the timestamp in the HTTP
header. I realized that the changes around the new 'encoded_obj' faux
Redis type meant it was now interpreting the timestamp as a string when
pulling it from Redis. I updated the api.simpleauth authentication class
so that it explicitly converts BOTH timestamps to floats when comparing
them and then explicitly stores the timestamp in Redis as a string. I
also did a little code cleanup.