Enable specifying component versions in .env
I'm looking at upgrading various components on our production servers,
and it occurred to me that we don't have to hardcode versions of Redis,
Python, Solr, etc. I've created .env variables for these to make it
easier to test and development against different versions without having
to commit that configuration to the repository. This will make it easier
for us to test upgrades to components, and it will make it easier for
anyone using or experimenting with the public version of this project to
test and develop against whatever versions of things they're using.

This includes the docker images (Python, MySQL/MariaDB, Postgres, Redis,
and Solr) and the 'luceneMatchVersion' in the solrconfig.xml file for
each Solr core.

Defaults matching most of the previous versions are in place so that
Docker will continue using the same versions if you do not specify your
own. The exception is the default Docker image used for Python in the
Dockerfile -- I've switched that from 3.10.12-buster to 3.10.12 (which
uses 'bookworm,' the latest Debian).

After applying this commit it's helpful to rebuild your Docker
environment.